requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-braces==1.8.1
django-cors-headers==1.1.0
django-filter==0.13.0
django-http-proxy==0.4.3
django-oauth-plus==2.2.8
django-oauth-toolkit==0.10.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.3.3
djangorestframework-oauth==1.1.0
feedparser==5.2.1
gunicorn==19.6.0
httplib2==0.9.2
Markdown==2.6.6
oauth2==1.9.0.post1
oauthlib==1.0.3
psycopg2==2.6.1
pytz==2016.4
requests==2.10.0
shortuuid==0.4.3
six==1.10.0
